# src/main/resources/openapi/api.yaml
openapi: 3.0.3
info:
  title: SpringTest API
  version: 1.0.0
paths:
...
  /authors/{id}:
    get:
      summary: Get author by id
      operationId: getAuthorById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Author found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthorResponse'
        '404':
          description: Author not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      summary: Update author
      operationId: updateAuthorById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthorUpdateRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthorResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Author not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Delete author
      operationId: deleteAuthorById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Deleted
        '404':
          description: Author not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /authors:
    get:
      summary: Get authors list
      operationId: getAuthors
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
      responses:
        '200':
          description: List of authors
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/AuthorResponse'
                  page:
                    type: integer
                  size:
                    type: integer
                  totalElements:
                    type: integer
                    format: int64
    post:
      summary: Create author
      operationId: createAuthor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthorCreateRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthorResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    AuthorResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        books:
          type: array
          items:
            $ref: '#/components/schemas/BookResponse'
    AuthorCreateRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 50
        books:
          type: array
          items:
            $ref: '#/components/schemas/BookCreateRequest'
    AuthorUpdateRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 50
        books:
          type: array
          maxItems: 200
          items:
            $ref: '#/components/schemas/BookUpdateRequest'
    BookResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
    BookCreateRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 100
    BookUpdateRequest:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          minLength: 1
          maxLength: 100
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        code:
          type: string
        timestamp:
          type: string
          format: date-time
        details:
          type: array
          items:
            type: string